requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
qdrant-client
pydantic
python-dotenv
langchain-qdrant
langchain-huggingface
sentence-transformers
pypdf
markdown
fastapi
uvicorn
nltk